refactor(backend): split index startup into startup modules
Move CORS and core Express middleware into startup/middleware.ts,
HTTP route mounting into startup/routes.ts, queue scheduler kickoff
into startup/workers.ts, and Node HTTP server creation into
startup/httpServer.ts. createApp in startup/appFactory.ts composes
middleware and routes so index.ts only validates config, builds the
app, listens, and logs.

Made-with: Cursor
